+++ services/__init__.py
@@ -1,6 +1,7 @@
 """Service layer initialization and exports."""
 
 from services.interfaces import MacroServiceInterface, SignalServiceInterface
+from services.macro_regime_service import MacroRegimeService
 from services.macro_service import MacroService
 from services.signal_service import SignalService
 
@@ -8,5 +9,6 @@ __all__ = [
     "MacroServiceInterface",
     "SignalServiceInterface",
     "MacroService",
+    "MacroRegimeService",
     "SignalService",
 ]